creacion maquetacion de panel de control para administrar la gestion de la app y un login para el mismo, prueba maquetacion de botones en profile me
