+++ lovely-frontend/src/app/app.component.css
@@ -1,50 +1,52 @@
 @import "../../node_modules/bootstrap/dist/css/bootstrap.min.css";
+.background-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(to bottom right, #eeeeee ,#f5f5f5 );
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;
+  z-index: -1;
+}
+:host {
+
+  min-height: 100vh;
+}
+
+.main-content {
+  flex: 1;
+  padding: 20px;
+  padding-bottom: 70px;
+  /*Altura del menú en móviles */
+}
+
+@media (min-width: 768px) {
   :host {
-    
-    min-height: 100vh;
+    flex-direction: row;
   }
-  
+
   .main-content {
-    flex: 1;
-    padding: 20px;
-    padding-bottom: 70px;  /*Altura del menú en móviles */
+    padding-left: 220px;
+    /* Ancho del menú + un poco de espacio extra */
+    padding-bottom: 20px;
   }
-  
-  @media (min-width: 768px) {
-    :host {
-      flex-direction: row;
-    }
-  
-    .main-content {
-      padding-left: 220px; /* Ancho del menú + un poco de espacio extra */
-      padding-bottom: 20px;
-    }
-  }
+}
 
-  .background-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: linear-gradient(to bottom right, #eeeeee ,#f5f5f5 );
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    z-index: -1;
-  }
-  
+
+main {
+  position: relative;
+  /*  z-index: 1; */
+  min-height: 100vh;
+  padding-bottom: 60px;
+  /* Ajusta esto según la altura de tu menú en móvil */
+}
+
+@media (min-width: 768px) {
   main {
-    position: relative;
-   /*  z-index: 1; */
-    min-height: 100vh;
-    padding-bottom: 60px; /* Ajusta esto según la altura de tu menú en móvil */
+    /*padding-left: 100px;  Ancho del menú en desktop */
+    padding-bottom: 0;
   }
-  
-  @media (min-width: 768px) {
-    main {
-      /*padding-left: 100px;  Ancho del menú en desktop */
-      padding-bottom: 0;
-    }
-  }
-
+}
