+++ lovely-frontend/src/app/app.component.html
@@ -1,5 +1,6 @@
 <div class="background-overlay"></div>
 <app-header></app-header>
-<main class="main-content">
   <router-outlet></router-outlet>
-</main>
+<!--<main class="main-content">
+ 
+</main>-->
