+++ lovely-frontend/src/app/components/admin/login-admin/login-admin.component.html
@@ -1,0 +1,20 @@
+<div class="main-container">
+    <div class="login-container">
+        <img src="../../../assets/lovely-full-dark.png" alt="Logo">
+        <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
+        <form>
+            <div class="form-floating">
+                <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
+                <label for="floatingInput">Email address</label>
+              </div>
+              <div class="form-floating my-3">
+                <input type="password" class="form-control" id="floatingPassword" placeholder="Password">
+                <label for="floatingPassword">Password</label>
+              </div>
+            <button class="btn btn-primary w-100 py-2" type="submit">Sign in</button>
+        </form>
+        <div class="footer-text">
+            &copy; 2024 Lovely
+        </div>
+    </div>
+</div>
