+++ lovely-frontend/src/app/components/admin/login-admin/login-admin.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-login-admin',
+  standalone: true,
+  imports: [],
+  templateUrl: './login-admin.component.html',
+  styleUrl: './login-admin.component.css'
+})
+export class LoginAdminComponent {
+
+}
