+++ lovely-frontend/src/app/components/layout/header/header.component.css
@@ -43,7 +43,7 @@ main {
   padding-bottom: 70px; /* Altura del menú en móviles */
 }
 .menu-container {
-  background: #f8f8f8;
+  background: #ffffff;
   box-shadow: 0 0 10px rgba(0,0,0,0.1);
 }
 .menu-item {
