+++ lovely-frontend/src/app/components/layout/header/header.component.html
@@ -24,6 +24,17 @@
         <span class="d-none d-md-block">Chat</span>
       </a>
     </div>
+    <div id="menu-item-4" class="menu-item rounded-3" routerLinkActive="active-btn" >
+      <a routerLink="/notification"  class="d-flex flex-column align-items-center position-relative"  >
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-bell-fill" viewBox="0 0 16 16">
+          <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2m.995-14.901a1 1 0 1 0-1.99 0A5 5 0 0 0 3 6c0 1.098-.5 6-2 7h14c-1.5-1-2-5.902-2-7 0-2.42-1.72-4.44-4.005-4.901"/>
+        </svg>
+        <span class="position-absolute top-0 start-10 translate-middle p-1 bg-danger border border-light rounded-circle">
+          <span class="visually-hidden">New alerts</span>
+        </span>
+        <span class="d-none d-md-block">Notificationes</span>
+      </a>
+    </div>
     <div id="menu-item-3" class="menu-item rounded-3" routerLinkActive="active-btn">
       <a routerLink="/perfil"  class="d-flex flex-column align-items-center" >
         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-person-lines-fill" viewBox="0 0 16 16">
@@ -31,31 +42,16 @@
         </svg>
         <span class="d-none d-md-block">Perfil</span>
       </a>
-      
       <div class="d-none d-md-block"></div>
     </div>
-      <div id="menu-item-4" class="menu-item rounded-3" routerLinkActive="active-btn" >
-        <a routerLink="/notification"  class="d-flex flex-column align-items-center position-relative"  >
-          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-bell-fill" viewBox="0 0 16 16">
-            <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2m.995-14.901a1 1 0 1 0-1.99 0A5 5 0 0 0 3 6c0 1.098-.5 6-2 7h14c-1.5-1-2-5.902-2-7 0-2.42-1.72-4.44-4.005-4.901"/>
-          </svg>
-          <span class="position-absolute top-0 start-10 translate-middle p-1 bg-danger border border-light rounded-circle">
-            <span class="visually-hidden">New alerts</span>
-          </span>
-          <span class="d-none d-md-block">Notificationes</span>
-        </a>
-      </div>
-   
-    
 
-
-<div class="menu-footer d-none d-md-block p-0">
-  <ul class="list-unstyled">
-    <li><a href="#">Privacidad</a> · <a href="#">Condiciones</a></li>
-    <li><a href="#">Publicidad</a> · <a href="#">Cookies</a></li>
-    <li><a href="#">Más</a></li>
-  </ul>
-  <p>Lovely © 2024</p>
-</div>
+    <div class="menu-footer d-none d-md-block p-0">
+      <ul class="list-unstyled">
+        <li><a href="#">Privacidad</a> · <a href="#">Condiciones</a></li>
+        <li><a href="#">Publicidad</a> · <a href="#">Cookies</a></li>
+        <li><a href="#">Más</a></li>
+      </ul>
+      <p>Lovely © 2024</p>
+    </div>
 </nav>
 }
