+++ lovely-frontend/src/app/components/login/login.component.html
@@ -1,5 +1,5 @@
 <nav class="nav nav-pills nav-justified">
-    <img src="../../../assets/lovely-full-dark.png" class="d-block mx-lg-auto img-fluid m-5" alt="logo" width="250" height="250" loading="lazy">
+    <img src="../../../assets/lovely-full-dark.png" class="d-block mx-auto img-fluid m-5" alt="logo" width="250" height="250" loading="lazy">
 </nav>
  <div class="container ">
     <div class="row ">
