+++ lovely-frontend/src/app/components/notification/notification.component.html
@@ -1,4 +1,5 @@
-<section class="container-fluid p-0">
+<main class="main-content">
+  <section class="container-fluid p-0">
     <!-- HEADER -->
     <div class="profile-header">
       <div class="container">
@@ -50,4 +51,6 @@
     </div>
 
  </section>
+
+</main>
