+++ lovely-frontend/src/app/components/perfil/perfil.component.css
@@ -12,6 +12,7 @@
     font-weight: normal;
     font-style: normal;
   }
+
 /* Estilos generales */
 body {
     font-family: 'OpenSans', sans-serif;
